+++ rocket/IMU/IMU.ino
@@ -6,8 +6,8 @@
 // http://www.ngdc.noaa.gov/geomag-web/#declination
 #define DECLINATION -12.6 // Declination (degrees) in Boulder, CO.
 
-float pitch, roll, yaw;
-float deltat;
+double pitch, roll, yaw;
+double deltat;
 
 void setup() {
   Serial.begin(9600);
@@ -20,16 +20,16 @@ void setup() {
 }
 
 void loop() {
-  float ax, ay, az, gx, gy, gz, mx, my, mz;
+  double ax, ay, az, gx, gy, gz, mx, my, mz;
   if (IMU.accelerationAvailable() && IMU.gyroscopeAvailable() && IMU.magneticFieldAvailable()) {
     IMU.readAcceleration(ax, ay, az);
     IMU.readGyroscope(gx, gy, gz);
     IMU.readMagneticField(mx, my, mz);
 
-    float roll = atan2(ay, az);
-    float pitch = atan2(-ax, sqrt(ay * ay + az * az));
+    double roll = atan2(ay, az);
+    double pitch = atan2(-ax, sqrt(ay * ay + az * az));
 
-    float heading;
+    double heading;
     if (my == 0) {
       heading = (mx < 0) ? PI : 0;
     } else {
